+++ src/language/language.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateLanguageDto } from './dto/create-language.dto';
 import { UpdateLanguageDto } from './dto/update-language.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -20,17 +20,24 @@ export class LanguageService {
     return this.languageRepository.find();
   }
 
-  findOne(id: number) {
-    return this.languageRepository.findOneBy({id});
-  }
+  async findOne(id: number) {
+    const languageFound = await this.languageRepository.findOneBy({ id });
+    if (!languageFound) throw new HttpException(`Language with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    return languageFound;
+}
 
-  async update(id: number, updateLanguageDto: UpdateLanguageDto) {
-    await this.languageRepository.update(id,updateLanguageDto)
+async update(id: number, updateLanguageDto: UpdateLanguageDto) {
+    const languageFound = await this.languageRepository.findOneBy({ id });
+    if (!languageFound) throw new HttpException(`Language with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    await this.languageRepository.update(id, updateLanguageDto);
     return updateLanguageDto;
-  }
+}
 
-  async remove(id: number) {
-    await this.languageRepository.delete(id)
+async remove(id: number) {
+    const languageFound = await this.languageRepository.findOneBy({ id });
+    if (!languageFound) throw new HttpException(`Language with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    await this.languageRepository.delete(id);
     return `The language with the id #${id} was deleted`;
-  }
+}
+
 }
