+++ src/skill/skill.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateSkillDto } from './dto/create-skill.dto';
 import { UpdateSkillDto } from './dto/update-skill.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -22,17 +22,26 @@ export class SkillService {
     return this.skillRepository.find();
   }
 
-  findOne(id: number) {
-    return this.skillRepository.findOneBy({id});
-  }
+  async findOne(id: number) {
+    const skillFound = await this.skillRepository.findOneBy({ id });
+    if (!skillFound) throw new HttpException(`Skill with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    return skillFound;
+}
 
-  async update(id: number, updateSkillDto: UpdateSkillDto) {
-    await this.skillRepository.update(id, updateSkillDto)
+async update(id: number, updateSkillDto: UpdateSkillDto) {
+    const skillFound = await this.skillRepository.findOneBy({ id });
+    if (!skillFound) throw new HttpException(`Skill with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    
+    await this.skillRepository.update(id, updateSkillDto);
     return updateSkillDto;
-  }
+}
 
-  async remove(id: number) {
-    await this.skillRepository.delete(id)
-    return `The education with the id #${id} was deleted`;
-  }
+async remove(id: number) {
+    const skillFound = await this.skillRepository.findOneBy({ id });
+    if (!skillFound) throw new HttpException(`Skill with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    
+    await this.skillRepository.delete(id);
+    return `The skill with the id #${id} was deleted`;
+}
+
 }
