+++ src/users/users.service.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -14,30 +14,41 @@ export class UsersService {
   }
 
   async create(createUserDto: CreateUserDto) {
-    
+    const userFound = await this.userRepository.findOne({
+      where: { email: createUserDto.email }});
+    if (!userFound) {
     const userCreated = this.userRepository.create(createUserDto);
     await this.userRepository.save(userCreated);
-
     return userCreated;
+    } else {
+      throw new HttpException(`User with Email ${createUserDto.email} already exists`, HttpStatus.BAD_REQUEST);
+    }
   }
 
   findAll() {
     return this.userRepository.find();
   }
 
-  findOne(id: number) {
-    return this.userRepository.findOneBy({id});
-  }
+  async findOne(id: number) {
+    const userFound = await this.userRepository.findOneBy({ id });
+    if (!userFound) throw new HttpException(`User with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    return userFound;
+}
 
-  async update(id: number, updateUserDto: UpdateUserDto) {
-
+async update(id: number, updateUserDto: UpdateUserDto) {
+    const userFound = await this.userRepository.findOneBy({ id });
+    if (!userFound) throw new HttpException(`User with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    
     await this.userRepository.update(id, updateUserDto);
-
     return updateUserDto;
-  }
+}
 
-  async remove(id: number) {
+async remove(id: number) {
+    const userFound = await this.userRepository.findOneBy({ id });
+    if (!userFound) throw new HttpException(`User with ID ${id} not found`, HttpStatus.NOT_FOUND);
+    
     await this.userRepository.delete(id);
-    return `This action removes a #${id} user`;
-  }
+    return `The user with the id #${id} was deleted`;
+}
+
 }
